Rename socket::raw to socket::tcp
Rename the raw module to tcp.  This name provides better meaning for the
transport implementation inside.
